Consolidate OpenFeature smoke specs
